requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
google-genai==0.2.2
gunicorn==21.2.0
